requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.4.0
Django==1.9.6
django-anymail==0.5
django-dotenv==1.4.1
django-extensions==1.6.7
django-mail-templated==2.6.5
enum34==1.1.6
gnureadline==6.3.8
idna==2.7
ipython==4.2.0
ipython-genutils==0.2.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
psycopg2==2.6.1
ptyprocess==0.6.0
requests==2.19.1
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wheel==0.24.0
